moisture_monitoring/requirements.txt
CherryPy==18.8.0
numpy==1.26.0
paho_mqtt==1.6.1
Requests==2.32.2
